+++ squirls-initialize/src/main/java/org/monarchinitiative/squirls/initialize/DatasourceProperties.java
@@ -83,4 +83,6 @@ public interface DatasourceProperties {
 
     int maxTranscriptSupportLevel();
 
+    boolean useNoncodingTranscripts();
+
 }